Various Fixes and Changes
I planned on using this addon for a few of the TTT servers I own. But
taking a look at the code I found a few errors that I decided to fix.
Fixed IsTTT being set as a local inside a if statement
Made Sure IsTTT is set after all the Gamemode files have been loaded.
Fixed Identifing Corpses being when they are removed.
Made it so that it announces bodies when they are id/removed.
Made it so it counts the bodies kills and ids those before its removed.
Cleaned up some other stuff.